moved files to seperate folders for component/page layout
