+++ index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="apple-touch-icon" href="./assets/images/fwf-question.png" type="image/png">
   <link rel="icon" href="./assets/images/fwf-question.png" type="image/png">
-  <link rel="stylesheet" href="./styles/index.css">
+  <link rel="stylesheet" href="./pages/LandingPage/styles/LandingPage.css">
   <title>Fun with flags</title>
 </head>
 
@@ -60,7 +60,7 @@
       </section>
     </form>
   </main>
-  <script src="./scripts/LandingPage.js" type="module"></script>
+  <script src="./pages/LandingPage/LangingPage.js" type="module"></script>
 </body>
 
 </html>
