+++ pages/FWFGame/index.html
@@ -5,9 +5,9 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <link rel="apple-touch-icon" href="./assets/images/fwf-question.png" type="image/png">
-  <link rel="icon" href="./assets/images/fwf-question.png" type="image/png">
+  <link rel="icon" href="../../assets/images/fwf-question.png" type="image/png">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="../styles/game.css">
+  <link rel="stylesheet" href="./styles/FWFGame.css">
   <title>Fun With Flags</title>
 </head>
 
@@ -35,12 +35,12 @@
     </article>
     <article class="fwf__game-over">
       <h2>Game Over</h2>
-      <img class="fwf__game-over-img" src="../assets/images/bazinga.png" alt="bazinga">
-      <a class="fwf__results-button" href="./results.html">View Results</a>
+      <img class="fwf__game-over-img" src="../../assets/images/bazinga.png" alt="bazinga">
+      <a class="fwf__results-button" href="../ResultsPage/index.html">View Results</a>
     </article>
   </main>
-  <script src="../scripts/FWFGame.js" type="module"></script>
-  <script src="../scripts/ProgressBarAnimation.js" type="module"></script>
+  <script src="./FWFGame.js" type="module"></script>
+  <script src="../../scripts/ProgressBarAnimation.js" type="module"></script>
 </body>
 
 </html>
