+++ pages/LandingPage/styles/LandingPage.css
@@ -213,4 +213,4 @@ main {
 .registration-form__submit:hover {
   background-color: #ffdf6c;
   color: #2e2e2e;
-}/*# sourceMappingURL=index.css.map */
+}/*# sourceMappingURL=LandingPage.css.map */
