+++ pages/ResultsPage/ResultsPage.js
@@ -1,4 +1,4 @@
-import { createPageElement } from "../utils/utils.js";
+import { createPageElement } from "../../utils/utils.js";
 
 
 async function buildResults() {
