+++ pages/ResultsPage/index.html
@@ -7,15 +7,15 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="apple-touch-icon" href="./assets/images/fwf-question.png" type="image/png">
   <link rel="icon" href="./assets/images/fwf-question.png" type="image/png">
-  <link rel="stylesheet" href="../styles/results.css">
+  <link rel="stylesheet" href="./styles/results.css">
   <title>Fun With Flags</title>
 </head>
 
 <body>
   <header class="header">
-    <a href="../index.html" class="header__logo"><img src="../assets/images/sheldon-cooper-presents-fun-w-flags.webp"
+    <a href="../../index.html" class="header__logo"><img src="../assets/images/sheldon-cooper-presents-fun-w-flags.webp"
         alt=""></a>
-    <a href="../index.html" class="header__play-again">Play Again?</a>
+    <a href="../../index.html" class="header__play-again">Play Again?</a>
   </header>
   <section class="results"></section>
   <section class="scoreboard">
@@ -35,7 +35,7 @@
   <footer class="footer">
     <div class="footer__sign">
       <p>Created by</p>
-      <p>Madhura Pendbhaje & Louis Flores</p>
+      <p>Madhura Pendbhaje & <a href="https://github.com/l0fl0">Louis Flores</a></p>
     </div>
   </footer>
   <script src="../scripts/ResultsPage.js" type="module"></script>
