+++ pages/ResultsPage/styles/results.scss
@@ -1,6 +1,6 @@
-@use "./partials/global";
-@use "./partials/mixins" as *;
-@use "./partials/variables" as *;
+@use "../../../partials/global";
+@use "../../../partials/mixins" as *;
+@use "../../../partials/variables" as *;
 
 .header {
 	flex-direction: column;
